requirements.txt
yfinance==0.2.65
pandas==2.1.4
numpy==1.24.3
matplotlib==3.8.2
plotly==5.17.0
streamlit==1.29.0
ta==0.10.2
scikit-learn==1.3.2
seaborn==0.13.0
requests==2.31.0
shioaji
python-dotenv 